% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{create_BYR_colors}
\alias{create_BYR_colors}
\title{Create 256 colors gradually transitioning from
Blue to Yellow to Red.}
\usage{
create_BYR_colors(num = 256L)
}
\arguments{
\item{num}{A integer number to specify the number of colors to generate. The default value is 256.}
}
\value{
A vector of RGB colors
}
\description{
Create 256 colors gradually transitioning from
Blue to Yellow to Red.
}
\examples{
colors = create_BYR_colors(101L)
require(graphics)
filled.contour(volcano,col=colors,nlevels=length(colors)-1,asp=1)
filled.contour(volcano,color.palette = create_BYR_colors, nlevels=256, asp = 1)
}
\author{
Jian Kang <jiankang@umich.edu>
}
