% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{create_img_coords}
\alias{create_img_coords}
\title{Create coordinates for images}
\usage{
create_img_coords(imgdim = c(50, 50), coord_range = c(-1, 1))
}
\arguments{
\item{imgdim}{integer vector for creating image dimensions}

\item{coord_range}{numeric vector for the range of coordinates for each dimension}
}
\value{
matrix of coordinates
}
\description{
Create coordinates for images
}
\examples{
coords_1 <- create_img_coords()
coords_2 <- create_img_coords(imgdim = c(10,20,30), coord_range=c(-1,1,-2,2,-3,3))
}
\author{
Jian Kang <jiankang@umich.edu>
}
