% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{gen_grids_df_4region}
\alias{gen_grids_df_4region}
\title{Generate grids for 4 regions}
\usage{
gen_grids_df_4region(side, d = 2)
}
\arguments{
\item{side}{side length of the square grid}

\item{d}{dimension}
}
\description{
Generate grids for 4 regions
}
