% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{generate_IonS_data}
\alias{generate_IonS_data}
\title{Simulate image on scalar regression data}
\usage{
generate_IonS_data(alpha_img, basis_sq, n, q = 2, sigma_M = 0.1)
}
\arguments{
\item{alpha_img}{the true image}

\item{basis_sq}{the basis object}

\item{n}{the sample size}

\item{q}{the number of confounders}
}
\description{
Simulate image on scalar regression data
}
