% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{generate_SonI_data}
\alias{generate_SonI_data}
\title{Simulate scalar on image regression data}
\usage{
generate_SonI_data(
  beta_img,
  n,
  q = 2,
  sigma_M = 0.1,
  sigma_C = 1,
  include_Confounder = T,
  sigma_Y = 0.1,
  SNR = NULL
)
}
\arguments{
\item{beta_img}{the true image}

\item{n}{the sample size}

\item{q}{the number of confounders}
}
\description{
Simulate scalar on image regression data
}
