% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Image_on_scalar}
\alias{Image_on_scalar}
\alias{IonS_CAVI_rho}
\title{Image on scalar regression}
\usage{
IonS_CAVI_rho(
  M,
  X,
  C,
  lambda,
  basis,
  rho,
  B,
  W,
  in_Sigma_inv,
  in_D,
  method,
  init_paras,
  sigmasq_step_controls,
  in_delta_rho,
  SGD_controls,
  initial_sigma_sq = 1,
  initial_tau_mu_sq = 0.01,
  initial_sigma_alpha_sq = 1,
  initial_sigma_zeta_sq = 1,
  initial_sigma_eta_sq = 1,
  eta_freq = 1L,
  max_iter = 1000L,
  mcmc_sample = 1000L,
  burnin = 1000L,
  thinning = 1L,
  begin_f_alpha = 0L,
  f_alpha_interval = 1L,
  paras_diff_tol = 1e-06,
  verbose = 5000L,
  save_profile = 1L,
  display_progress = FALSE
)
}
\arguments{
\item{M}{Matrix of functional images}

\item{X}{vector covariate}

\item{lambda}{thresholding parameter}

\item{rho}{scaling parameter in CAR model}

\item{B}{matrix of covariance neighborhood}

\item{in_Sigma_inv}{inverse of the covariance matrix obtained from CAR model}

\item{in_D}{diagonal vector, obtained from CAR model}

\item{method}{method for the optimization: "CAVI" or "SGD"(SSVI in the paper)}

\item{init_paras}{initial values for the parameters}

\item{y}{outcome}

\item{batch_control}{control parameters for the batch algorithm}
}
\value{
A List object with the following component
\itemize{
\item post_mean posterior mean of the parameters
\item iter total number of iterations
\item trace trace of each parameter
\item vb_control control variables
\item test_output to output temp variables, for testing only
\item elapsed time used
}
}
\description{
Scalar on Image regression using the sparse-mean prior
}
