% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{neighbor_sparse_mat_RANN}
\alias{neighbor_sparse_mat_RANN}
\title{Generate neighborhood matrix for the CAR model}
\usage{
neighbor_sparse_mat_RANN(
  grids,
  scale = 2,
  rho = 0.3,
  bandwidth = NULL,
  if_beep = F
)
}
\arguments{
\item{grids}{the location of the data}

\item{scale}{scale parameter to resize the distance}

\item{rho}{the correlation parameter}

\item{bandwidth}{the bandwidth of the neighborhood}
}
\description{
Generate neighborhood matrix for the CAR model
}
