% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{plot_2D_funcs}
\alias{plot_2D_funcs}
\title{Draw two-dimensional functions}
\usage{
plot_2D_funcs(f, grids, names = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{f}{numeric matrix with each column representing the function values evaluated at grids}

\item{grids}{numeric matrix with number grid points}

\item{names}{character vector where the length equals to the number of columns of f}

\item{nrow}{number of rows in the layout for more than one functions}

\item{ncol}{number of columns in the layout for more than one functions}
}
\description{
Draw two-dimensional functions
}
\examples{
n_sqrt <- 50
x <- as.matrix(expand.grid(seq(-1,1,length=n_sqrt),seq(-1,1,length=n_sqrt)))
bases <- create_random_ReLU_bases(x,10)
colnames(bases$f) <- paste("basis",1:ncol(bases$f))
L <- ncol(bases$f)
m <- 5
theta <- matrix(rnorm(L*m),nrow=L,ncol=m)
y <- bases$f\%*\%theta
g <- plot_2D_funcs(y,bases$x)
plot(g)
}
\author{
Jian Kang <jiankang@umich.edu>
}
