% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{plot_img}
\alias{plot_img}
\title{plot_img}
\usage{
plot_img(img, grids_df, title = "img", col_bar = NULL, legend = T)
}
\arguments{
\item{img}{A vector of input image}

\item{grids_df}{A data frame to indicate the position of pixels}
}
\description{
plot_img
}
