% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{simul_images}
\alias{simul_images}
\title{Simulate multiple images}
\usage{
simul_images(
  num_imgs = 9,
  coords = NULL,
  bases = NULL,
  imgdim = c(50, 50),
  coord_range = c(-1, 1),
  max_num_bases = 500,
  rescale = TRUE
)
}
\arguments{
\item{num_imgs}{positive integer for number of images}

\item{coords}{numeric matrix}

\item{bases}{numeric matrix}

\item{imgdim}{integer vector}

\item{coord_range}{numeric vector}

\item{rescale}{logical scalar (TRUE/FALSE) indicating whether or not rescale the}
}
\value{
list object with three elements 
\describe{
\item{img}{numeric matrix of multiple images}
\item{coords}{numeric matrix for coordinates}
\item{bases}{list object with basis functions (f), coords (x) and other information}
\item{theta}{numeric matrix of coefficients to generate images}
}
}
\description{
Simulate multiple images
}
\examples{
imgdat <- simul_images(1000)
plot_2D_funcs(imgdat$img[,1:10],dat$bases$x)
}
\author{
Jian Kang <jiankang@umich.edu>
}
