% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{simul_linear_reg}
\alias{simul_linear_reg}
\title{Simulate response variables from linear regression}
\usage{
simul_linear_reg(
  X,
  betacoef = NULL,
  R_sq = 0.9,
  sigma_sq = NULL,
  intercept = 0
)
}
\arguments{
\item{X}{design matrix}

\item{betacoef}{numeric vector for regression coefficients}

\item{R_sq}{R-squared indicating the signal-to-noise ratio}

\item{sigma_sq}{noise variance}

\item{intercept}{numeric scalar to specify the intercept. The default value is 0.0.}
}
\value{
List object with multiple elements
\describe{
\item{y}{vector of the response variable}
\item{X}{covariate matrix}
\item{betacoef}{vector of regression coefficients}
\item{R_sq}{R squared}
\item{sigma_sq}{Random noise variance}
\item{intercept}{scalar intercept}
}
}
\description{
Simulate response variables from linear regression
}
\examples{
n <- 100
imgs <- simul_images(n,imgdim=c(50,50))
coords <- create_img_coords(imgdim=c(50,50))
signals <- spatial_sparse_signals(coords,p=1L,shape="UMlogo")
regdat <- simul_linear_reg(X=t(imgs$img),betacoef=signals$img/sqrt(nrow(coords)))
fig1 <- ggplot2::qplot(regdat$y,bins=round(sqrt(length(regdat$y))))
fig2 <- with(regdat,ggplot2::qplot(X\%*\%betacoef,y))
fig3 <- plot_2D_funcs(f=regdat$betacoef,grids=coords)
fig4 <- plot_2D_funcs(f=imgs$img[,1],grids=coords,names="X1")
gridExtra::grid.arrange(fig1,fig2,fig3,fig4,nrow=2,ncol=2)
}
\author{
Jian Kang <jiankang@umich.edu>
}
