% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{simulate_round_image}
\alias{simulate_round_image}
\title{simulate_round_image}
\usage{
simulate_round_image(
  center_shift = c(0, 0),
  lambda = 0.1,
  side = 30,
  range = c(0, 1)
)
}
\description{
simulate_round_image
}
