% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{simulate_triang_image}
\alias{simulate_triang_image}
\title{simulate_triang_image}
\usage{
simulate_triang_image(
  center_shift = c(0, 0),
  side = 30,
  rad = 0.5,
  radiance_bool = T,
  range = c(0, 1)
)
}
\description{
simulate_triang_image
}
