% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_image_func.R
\name{spatial_sparse_signals}
\alias{spatial_sparse_signals}
\title{Generate one or multiple 2D images with sparse signals with different shapes}
\usage{
spatial_sparse_signals(
  coords,
  p = 1L,
  shape = c("circle", "triangle", "square", "Tshape", "UMlogo"),
  offset = 0.2,
  center = c(0, 0),
  effect_size = 1,
  smooth = 0.1,
  radius = 0.5
)
}
\arguments{
\item{coords}{n by 2 matrix for (x y) coordinates}

\item{p}{positive integer indicating the number of images}

\item{shape}{character scalar or vector including one or more of the following shapes "circle","triangle","square","Tshape" and "UMlogo"}

\item{offset}{minimum signal value in range(0,1)}

\item{effect_size}{positive numeric scalar indicating the magnitude of nonzero signals}

\item{smooth}{positive numeric scalar indicating the smoothness of the nonzero signals}

\item{radius}{positive numeric scalar indicating the size of activation shape in the image}
}
\description{
Generate one or multiple 2D images with sparse signals with different shapes
}
\examples{
coords <- create_img_coords(imgdim = c(32,32)) 
signals <- spatial_sparse_signals(coords,p=5L,shape=c("circle","square","triangle","Tshape","UMlogo"),radius=0.5,
center = c(0.0,0.0), offset=0.1, effect_size=2.0,smooth=1)
plot_2D_funcs(signals$img,signals$coords)
write.csv(signals$coords, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/coords_32x32_eff2_offset01.csv",row.names = F)
write.csv(signals$img, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/image_32x32_eff2_offset01.csv",row.names = F)

coords <- create_img_coords(imgdim = c(64,64)) 
signals <- spatial_sparse_signals(coords,p=5L,shape=c("circle","square","triangle","Tshape","UMlogo"),radius=0.5,
center = c(0.0,0.0), offset=0.1, effect_size=2.0,smooth=1)
plot_2D_funcs(signals$img,signals$coords)
write.csv(signals$coords, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/coords_64x64_eff2_offset01.csv",row.names = F)
write.csv(signals$img, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/image_64x64_eff2_offset01.csv",row.names = F)

coords <- create_img_coords(imgdim = c(128,128)) 
signals <- spatial_sparse_signals(coords,p=5L,shape=c("circle","square","triangle","Tshape","UMlogo"),radius=0.5,
center = c(0.0,0.0), offset=0.1, effect_size=2.0,smooth=1)
plot_2D_funcs(signals$img,signals$coords)
write.csv(signals$coords, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/coords_128x128_eff2_offset01.csv",row.names = F)
write.csv(signals$img, "~/Dropbox (University of Michigan)/DeepLearningGPKernel/Code/DKL/I_S/sim/data/image_128x128_eff2_offset01.csv",row.names = F)
}
\author{
Jian Kang <jiankang@umich.edu>
}
