% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Scalar_on_Image}
\alias{Scalar_on_Image}
\title{Scalar on Image regression standalone}
\arguments{
\item{y}{outcome}

\item{X}{vector covariate}

\item{M}{Matrix of functional images}

\item{lambda}{thresholding parameter}

\item{rho}{scaling parameter in CAR model}

\item{B}{matrix of covariance neighborhood}
}
\description{
Scalar on Image regression using the sparse-mean prior
}
